Only show Lock Profile button in Profile settings tab (not all tabs)
