+++ .changelog/1850.bugfix.md
@@ -1,0 +1,1 @@
+Update buttons layout in mobile settings and fix double scrollbar in extension
