+++ src/app/components/Toolbar/Features/Profile/index.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react'
-import { useSelector } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 import { useNavigate } from 'react-router-dom'
 import { Trans, useTranslation } from 'react-i18next'
 import { Box } from 'grommet/es6/components/Box'
@@ -9,6 +9,10 @@ import { User } from 'grommet-icons/es6/icons/User'
 import { selectUnlockedStatus } from 'app/state/selectUnlockedStatus'
 import { UpdatePassword } from './UpdatePassword'
 import { DeleteProfileButton } from '../../../Persist/DeleteProfileButton'
+import { Lock } from 'grommet-icons/es6/icons/Lock'
+import { Logout } from 'grommet-icons/es6/icons/Logout'
+import { selectIsLockableOrCloseable } from 'app/state/selectIsLockableOrCloseable'
+import { persistActions } from 'app/state/persist'
 
 type ProfileEmptyStateProps = {
   children: ReactNode
@@ -28,41 +32,72 @@ interface ProfileProps {
 export const Profile = ({ closeHandler }: ProfileProps) => {
   const { t } = useTranslation()
   const unlockedStatus = useSelector(selectUnlockedStatus)
-  const isAvailable = unlockedStatus === 'unlockedProfile'
+  const isProfileAvailable = unlockedStatus === 'unlockedProfile'
   const navigate = useNavigate()
-
-  if (!isAvailable) {
-    return (
-      <ProfileEmptyState>
-        <Box style={{ display: 'block' }}>
-          <Trans
-            i18nKey="toolbar.profile.notAvailable"
-            t={t}
-            components={{
-              OpenWalletButton: (
-                <Button
-                  color="link"
-                  onClick={() => {
-                    closeHandler()
-                    navigate('/open-wallet')
-                  }}
-                />
-              ),
-            }}
-            defaults="You can setup your profile while <OpenWalletButton>opening a wallet</OpenWalletButton>."
-          />
-        </Box>
-      </ProfileEmptyState>
-    )
+  const dispatch = useDispatch()
+  const isLockableOrCloseable = useSelector(selectIsLockableOrCloseable)
+  const closeWallet = () => {
+    navigate('/')
+    dispatch(persistActions.lockAsync())
+  }
+  const lockProfile = () => {
+    dispatch(persistActions.lockAsync())
   }
 
   return (
-    <Box flex="grow">
-      <UpdatePassword />
-      <Box gap="small" margin={{ top: 'medium' }} alignSelf="start">
-        <Text>{t('toolbar.profile.deletion', 'Deletion')}</Text>
-        <DeleteProfileButton prominent />
+    <>
+      <Box flex="grow">
+        {!isProfileAvailable && (
+          <ProfileEmptyState>
+            <Box style={{ display: 'block' }}>
+              <Trans
+                i18nKey="toolbar.profile.notAvailable"
+                t={t}
+                components={{
+                  OpenWalletButton: (
+                    <Button
+                      color="link"
+                      onClick={() => {
+                        closeHandler()
+                        navigate('/open-wallet')
+                      }}
+                    />
+                  ),
+                }}
+                defaults="You can setup your profile while <OpenWalletButton>opening a wallet</OpenWalletButton>."
+              />
+            </Box>
+          </ProfileEmptyState>
+        )}
+
+        {isProfileAvailable && (
+          <>
+            <UpdatePassword />
+            <Box gap="small" margin={{ top: 'medium' }} alignSelf="start">
+              <Text>{t('toolbar.profile.deletion', 'Deletion')}</Text>
+              <DeleteProfileButton prominent />
+            </Box>
+          </>
+        )}
       </Box>
-    </Box>
+      <Box align="end" margin={{ top: 'large' }}>
+        {isLockableOrCloseable === 'closeable' && (
+          <Button
+            data-testid="profile-modal-close-wallet"
+            icon={<Logout />}
+            label={t('menu.closeWallet', 'Close wallet')}
+            onClick={() => closeWallet()}
+          />
+        )}
+        {isLockableOrCloseable === 'lockable' && (
+          <Button
+            data-testid="profile-modal-lock-wallet"
+            icon={<Lock />}
+            label={t('menu.lockProfile', 'Lock profile')}
+            onClick={() => lockProfile()}
+          />
+        )}
+      </Box>
+    </>
   )
 }
