+++ src/app/components/Toolbar/Features/ProfileModalButton/index.tsx
@@ -3,17 +3,11 @@
  * ProfileModalButton
  *
  */
-import { useNavigate } from 'react-router-dom'
-import { Box } from 'grommet/es6/components/Box'
 import { Button } from 'grommet/es6/components/Button'
 import { ResponsiveContext } from 'grommet/es6/contexts/ResponsiveContext'
-import { Lock } from 'grommet-icons/es6/icons/Lock'
-import { Logout } from 'grommet-icons/es6/icons/Logout'
 import React, { memo, useState } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector } from 'react-redux'
 import { selectAddress, selectHasAccounts } from 'app/state/wallet/selectors'
-import { selectIsLockableOrCloseable } from 'app/state/selectIsLockableOrCloseable'
-import { persistActions } from 'app/state/persist'
 import { AccountSelector } from '../AccountSelector'
 import { JazzIcon } from '../../../JazzIcon'
 import { sidebarSmallSizeLogo, sidebarMediumSizeLogo } from '../../../../../styles/theme/elementSizes'
@@ -26,26 +20,14 @@ import { Contacts } from '../Contacts'
 import { Profile } from '../Profile'
 import { Settings } from '../Settings'
 import { LayerContainer } from '../LayerContainer'
-import { ButtonLink } from '../../../ButtonLink'
-import { Add } from 'grommet-icons/es6/icons/Add'
 
 export const ProfileModalButton = memo(() => {
   const { t } = useTranslation()
-  const navigate = useNavigate()
-  const dispatch = useDispatch()
-  const isLockableOrCloseable = useSelector(selectIsLockableOrCloseable)
   const walletHasAccounts = useSelector(selectHasAccounts)
   const address = useSelector(selectAddress)
   const [layerVisibility, setLayerVisibility] = useState(false)
   const isMobile = React.useContext(ResponsiveContext) === 'small'
   const hideLayer = () => setLayerVisibility(false)
-  const closeWallet = () => {
-    navigate('/')
-    dispatch(persistActions.lockAsync())
-  }
-  const lockProfile = () => {
-    dispatch(persistActions.lockAsync())
-  }
 
   if (!walletHasAccounts) {
     return null
@@ -79,26 +61,6 @@ export const ProfileModalButton = memo(() => {
               <Settings />
             </Tab>
           </Tabs>
-          {isMobile && (
-            <Box direction="row" justify="between" margin={{ top: 'large', bottom: 'medium' }} gap="medium">
-              {isLockableOrCloseable === 'closeable' && (
-                <Button
-                  data-testid="profile-modal-close-wallet"
-                  icon={<Logout />}
-                  label={t('menu.closeWallet', 'Close wallet')}
-                  onClick={() => closeWallet()}
-                />
-              )}
-              {isLockableOrCloseable === 'lockable' && (
-                <Button
-                  data-testid="profile-modal-lock-wallet"
-                  icon={<Lock />}
-                  label={t('menu.lockProfile', 'Lock profile')}
-                  onClick={() => lockProfile()}
-                />
-              )}
-            </Box>
-          )}
         </LayerContainer>
       )}
     </>
